Added constructor for custom notification length. Redesigned all commands using notifications instead of chat. Added a hide command and fixed some mistakes with other commands.

+++ src/main/java/dev/revere/virago/client/commands/FriendCommand.java
@@ -3,8 +3,10 @@ package dev.revere.virago.client.commands;
 import dev.revere.virago.Virago;
 import dev.revere.virago.api.command.AbstractCommand;
 import dev.revere.virago.api.command.CommandData;
+import dev.revere.virago.client.notification.NotificationType;
 import dev.revere.virago.client.services.FriendService;
 import dev.revere.virago.client.services.ModuleService;
+import dev.revere.virago.client.services.NotificationService;
 import dev.revere.virago.util.Logger;
 import net.optifine.Log;
 
@@ -13,50 +15,66 @@ import net.optifine.Log;
  * @project Virago
  * @date 3/21/2024
  */
-@CommandData(aliases = {"friend"}, description = "Toggle a module", syntax = "friend <player>")
+@CommandData(aliases = {"friend"}, description = "Toggle a module", syntax = "<add/remove/list/clear> <player>")
 public class FriendCommand extends AbstractCommand {
 
     @Override
     public void executeCommand(String line, String[] args) {
         FriendService service = Virago.getInstance().getServiceManager().getService(FriendService.class);
+        NotificationService notificationService = Virago.getInstance().getServiceManager().getService(NotificationService.class);
+
+        if (args.length < 2) {
+            notificationService.notify(NotificationType.NO, "Command Manager", "Please provide an option! " + getSyntax());
+            return;
+        }
+
+        if (args.length < 3 && !"list".equals(args[1]) && !args[1].equalsIgnoreCase("clear")) {
+            if ("remove".equals(args[1]) || "add".equals(args[1])) {
+                notificationService.notify(NotificationType.NO, "Command Manager", "Please provide a player!");
+                return;
+            }
+            return;
+        }
 
         String option = args[1].toLowerCase();
 
         switch(option) {
             case "remove": {
                 String player = args[2];
-
                 if(!service.isFriend(player)) {
-                    Logger.addChatMessage("You do not have " + player + " as a friend.");
+                    notificationService.notify(NotificationType.NO, "Command Manager", "You do not have " + player + " as a friend.");
                     break;
                 }
 
                 service.removeFriend(player);
-                Logger.addChatMessage("You have removed " + player + " as a friend.");
+                notificationService.notify(NotificationType.YES, "Command Manager", "You have removed " + player + " from your friend list.");
                 break;
             }
 
             case "add": {
                 String player = args[2];
-
                 if(service.isFriend(player)) {
-                    Logger.addChatMessage("You already have " + player + " as a friend.");
+                    notificationService.notify(NotificationType.NO, "Command Manager", "You already have " + player + " as a friend.");
                     break;
                 }
 
                 service.addFriend(player);
-                Logger.addChatMessage("You have added " + player + " as a friend.");
+                notificationService.notify(NotificationType.YES, "Command Manager", "You have added " + player + " as a friend.");
                 break;
             }
 
             case "list": {
-                Logger.addChatMessage("Friends List: " + service.getFriends().toString().replace("[", "").replace("]", ""));
+                if (service.getFriends().isEmpty()) {
+                    notificationService.notify(NotificationType.NO, "Command Manager", "You do not have any friends.");
+                    break;
+                }
+                Logger.addChatMessage("Friends: " + service.getFriends().toString().replace("[", "").replace("]", ""));
                 break;
             }
 
             case "clear": {
                 service.getFriends().clear();
-                Logger.addChatMessage("You have cleared your friends list.");
+                notificationService.notify(NotificationType.YES, "Command Manager", "You have cleared your friends list.");
                 break;
             }
         }
